Some more issues from renaming
